Time - The Agency That's Got Your Back - CFPB
Preview: A little-known government watchdog is striking fear into the
lending industry

The first time Deborah Jacobs heard about the Consumer Financial
Protection Bureau, it struck her as a terrible idea. It was 2010, and
Jacobs, a 65-year-old retired teacher, was one year into the mortgage on
her new house in Brooklyn, Mich. ...